requirements.txt
paho-mqtt==1.6.1
cryptography==41.0.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
joblib==1.3.0 